refact structrue
