+++ src/__tests__/App.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import App from '../routes/App';
+
+it('App', () => {
+  const div = document.createElement('div');
+  ReactDOM.render(<App />, div);
+});
