+++ src/__tests__/HomeView.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import HomeView from '../components/HomeView';
+
+it('HomeView renders without crashing', () => {
+  const div = document.createElement('div');
+  ReactDOM.render(<HomeView />, div);
+});
