+++ src/assets/img/sky.jpg
