+++ src/assets/sass/index.scss
@@ -1,0 +1,20 @@
+// Some best-practice CSS that's useful for most apps
+// Just remove them if they're not what you want
+@import './layout';
+
+html {
+  box-sizing: border-box;
+}
+
+html,
+body {
+  margin: 0;
+  padding: 0;
+  height: 100%;
+}
+
+*,
+*:before,
+*:after {
+  box-sizing: inherit;
+}
