+++ src/components/HomeView.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import sky from '../assets/img/sky.jpg';
+
+function HomeView() {
+  return (
+    <div>
+      <img alt="sky" className="bg img-fluid" src={sky} />
+    </div>
+  );
+}
+
+export default HomeView;
