--- src/components/__tests__/App.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from '../App';
-
-it('App', () => {
-  const div = document.createElement('div');
-  ReactDOM.render(<App />, div);
-});
