--- src/components/__tests__/PrivateRoute.js
@@ -1,19 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { BrowserRouter, Route } from 'react-router-dom';
-import PrivateRoute from '../PrivateRoute';
-
-describe('(Component) PrivateRoute', () => {
-  it('isAuthenticated', () => {
-    const div = document.createElement('div');
-    function Component() {
-      return <div>hello</div>;
-    }
-    ReactDOM.render(
-      <BrowserRouter>
-        <PrivateRoute isAuthenticated component={Component} />
-      </BrowserRouter>,
-      div
-    );
-  });
-});
