+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import 'bootstrap/dist/css/bootstrap.css';
-import './styles/index.css';
-import App from './components/App';
+import './assets/sass/index.css';
+import App from './routes/App';
 import registerServiceWorker from './utils/registerServiceWorker';
 
 ReactDOM.render(<App />, document.getElementById('root'));
