--- src/layouts/index.js
@@ -1,3 +1,0 @@
-import Layout from './containers/Layout'
-
-export default Layout
