--- src/routes/Home/components/HomeView.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import sky from '../assets/sky.jpg';
-
-function HomeView() {
-  return (
-    <div>
-      <img alt="sky" className="bg img-fluid" src={sky} />
-    </div>
-  );
-}
-
-export default HomeView;
