--- src/routes/Home/index.js
@@ -1,4 +1,0 @@
-import HomeView from './components/HomeView'
-
-// Sync route definition
-export default HomeView
