--- src/routes/Login/index.js
@@ -1,3 +1,0 @@
-import Login from './containers/Login'
-
-export default Login
