--- src/routes/Zen/containers/ZenContainer.js
@@ -1,15 +1,0 @@
-import { connect } from 'react-redux'
-import { fetchZen, clearZen } from '../modules/zen'
-
-import Zen from '../components/Zen'
-
-const mapDispatchToProps = {
-  fetchZen,
-  clearZen
-}
-
-const mapStateToProps = (state) => ({
-  zen: state.zen
-})
-
-export default connect(mapStateToProps, mapDispatchToProps)(Zen)
