--- src/utils/__tests__/registerServiceWorker.js
@@ -1,8 +1,0 @@
-import register, { unregister } from '../registerServiceWorker';
-
-describe('registerServiceWorker.js', () => {
-  it('not throw', () => {
-    expect(register).not.toThrow();
-    expect(unregister).not.toThrow();
-  });
-});
